[api backend] Allow per package build flag in project meta
Add a new feature to allow per package build flags in project meta.
This is useful specially in the case of project links where it is
not possible to enable / disable packages because they don't really
exist in the linked from project.

An example of using this feature is adding the following line to a
project's meta:

<build><disable arch="i586" repository="i586" package="acl"/></build>

Signed-off-by: Islam Amer <pharon@gmail.com>